Change to use a boolean option for column selections.
